Update to mojo fd8d18dbf7a4f9ed8a607648cca6d2c595b1c580 (#2770)
* Update to mojo fd8d18dbf7a4f9ed8a607648cca6d2c595b1c580

* Update mojo_sdk

* Update to account for Mojo API changes

* Backport to older base

+++ mojo/edk/system/async_waiter.h
@@ -25,7 +25,7 @@ class AsyncWaiter final : public Awakable {
 
  private:
   // |Awakable| implementation:
-  bool Awake(MojoResult result, uintptr_t context) override;
+  bool Awake(MojoResult result, uint64_t context) override;
 
   AwakeCallback callback_;
