+++ mojo/edk/system/waiter_test_utils.h
@@ -49,7 +49,7 @@ class SimpleWaiterThread : public test::SimpleTestThread {
  public:
   // For the duration of the lifetime of this object, |*result| belongs to it
   // (in the sense that it will write to it whenever it wants).
-  SimpleWaiterThread(MojoResult* result, uint32_t* context);
+  SimpleWaiterThread(MojoResult* result, uint64_t* context);
   ~SimpleWaiterThread() override;  // Joins the thread.
 
   Waiter* waiter() { return &waiter_; }
@@ -58,7 +58,7 @@ class SimpleWaiterThread : public test::SimpleTestThread {
   void Run() override;
 
   MojoResult* const result_;
-  uint32_t* const context_;
+  uint64_t* const context_;
   Waiter waiter_;
 
   MOJO_DISALLOW_COPY_AND_ASSIGN(SimpleWaiterThread);
@@ -75,10 +75,10 @@ class WaiterThread : public test::SimpleTestThread {
   WaiterThread(util::RefPtr<Dispatcher>&& dispatcher,
                MojoHandleSignals handle_signals,
                MojoDeadline deadline,
-               uint32_t context,
+               uint64_t context,
                bool* did_wait_out,
                MojoResult* result_out,
-               uint32_t* context_out,
+               uint64_t* context_out,
                HandleSignalsState* signals_state_out);
   ~WaiterThread() override;
 
@@ -88,10 +88,10 @@ class WaiterThread : public test::SimpleTestThread {
   const util::RefPtr<Dispatcher> dispatcher_;
   const MojoHandleSignals handle_signals_;
   const MojoDeadline deadline_;
-  const uint32_t context_;
+  const uint64_t context_;
   bool* const did_wait_out_;
   MojoResult* const result_out_;
-  uint32_t* const context_out_;
+  uint64_t* const context_out_;
   HandleSignalsState* const signals_state_out_;
 
   Waiter waiter_;
