+++ mojo/edk/system/waiter_unittest.cc
@@ -36,14 +36,14 @@ class WaitingThread : public test::SimpleTestThread {
       : deadline_(deadline),
         done_(false),
         result_(MOJO_RESULT_UNKNOWN),
-        context_(static_cast<uint32_t>(-1)) {
+        context_(static_cast<uint64_t>(-1)) {
     waiter_.Init();
   }
 
   ~WaitingThread() override { Join(); }
 
   void WaitUntilDone(MojoResult* result,
-                     uint32_t* context,
+                     uint64_t* context,
                      MojoDeadline* elapsed) {
     for (;;) {
       {
@@ -66,7 +66,7 @@ class WaitingThread : public test::SimpleTestThread {
   void Run() override {
     Stopwatch stopwatch;
     MojoResult result;
-    uint32_t context = static_cast<uint32_t>(-1);
+    uint64_t context = static_cast<uint64_t>(-1);
     MojoDeadline elapsed;
 
     stopwatch.Start();
@@ -88,7 +88,7 @@ class WaitingThread : public test::SimpleTestThread {
   Mutex mutex_;
   bool done_ MOJO_GUARDED_BY(mutex_);
   MojoResult result_ MOJO_GUARDED_BY(mutex_);
-  uint32_t context_ MOJO_GUARDED_BY(mutex_);
+  uint64_t context_ MOJO_GUARDED_BY(mutex_);
   MojoDeadline elapsed_ MOJO_GUARDED_BY(mutex_);
 
   MOJO_DISALLOW_COPY_AND_ASSIGN(WaitingThread);
@@ -96,7 +96,7 @@ class WaitingThread : public test::SimpleTestThread {
 
 TEST(WaiterTest, Basic) {
   MojoResult result;
-  uint32_t context;
+  uint64_t context;
   MojoDeadline elapsed;
 
   // Finite deadline.
@@ -155,7 +155,7 @@ TEST(WaiterTest, Basic) {
     thread.Start();
     thread.WaitUntilDone(&result, &context, &elapsed);
     EXPECT_EQ(MOJO_RESULT_DEADLINE_EXCEEDED, result);
-    EXPECT_EQ(static_cast<uint32_t>(-1), context);
+    EXPECT_EQ(static_cast<uint64_t>(-1), context);
     EXPECT_GT(elapsed, (2 - 1) * test::EpsilonTimeout());
     EXPECT_LT(elapsed, (2 + 1) * test::EpsilonTimeout());
   }
@@ -216,7 +216,7 @@ TEST(WaiterTest, TimeOut) {
   MojoDeadline elapsed;
 
   Waiter waiter;
-  uint32_t context = 123;
+  uint64_t context = 123;
 
   waiter.Init();
   stopwatch.Start();
@@ -247,7 +247,7 @@ TEST(WaiterTest, TimeOut) {
 // The first |Awake()| should always win.
 TEST(WaiterTest, MultipleAwakes) {
   MojoResult result;
-  uint32_t context;
+  uint64_t context;
   MojoDeadline elapsed;
 
   {
